Add support to the decimal module
The idea is that float values are converted to Decimal as soon as they
arrive. After that, they are kept and manipulated with the best
precision possible, not limiting the number of decimal places until it
is necessary. The only instance where this happens is when a new order
is places.